clawdbot/scraper/requirements.txt
requests>=2.31.0
pyyaml>=6.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
